rename input overlay to inline menu
