+++ .github/workflows/test-overlay-autofill.yml
@@ -1,5 +1,5 @@
-name: test-overlay
-run-name: Overlay tests with extension build ${{ inputs.CLIENTS_BRANCH }} by @${{ github.actor }}
+name: test-inline-menu
+run-name: Inline menu tests with extension build ${{ inputs.CLIENTS_BRANCH }} by @${{ github.actor }}
 on:
   workflow_dispatch:
     inputs:
@@ -84,7 +84,7 @@ jobs:
           npm run setup:test-site
 
       - name: Run autofill tests
-        run: npm run test:static:overlay:headless
+        run: npm run test:static:inline-menu:headless
 
       - name: Upload results as artifact
         uses: actions/upload-artifact@694cdabd8bdb0f10b2cea11669e1bf5453eed0a6 # v4.2.0
