+++ package.json
@@ -60,8 +60,8 @@
     "test:static:autofill:headless": "npm run pretest && NODE_EXTRA_CA_CERTS=ssl.crt TARGET=static HEADLESS=true playwright test tests/autofill-forms.spec.ts",
     "test:static:notification": "npm run pretest && NODE_EXTRA_CA_CERTS=ssl.crt TARGET=static playwright test tests/notification-bar.spec.ts",
     "test:static:notification:headless": "npm run pretest && NODE_EXTRA_CA_CERTS=ssl.crt TARGET=static HEADLESS=true playwright test tests/notification-bar.spec.ts",
-    "test:static:overlay": "npm run pretest && NODE_EXTRA_CA_CERTS=ssl.crt TARGET=static playwright test tests/input-overlay.spec.ts",
-    "test:static:overlay:headless": "npm run pretest && NODE_EXTRA_CA_CERTS=ssl.crt TARGET=static HEADLESS=true playwright test tests/input-overlay.spec.ts",
+    "test:static:inline-menu": "npm run pretest && NODE_EXTRA_CA_CERTS=ssl.crt TARGET=static playwright test tests/inline-menu.spec.ts",
+    "test:static:inline-menu:headless": "npm run pretest && NODE_EXTRA_CA_CERTS=ssl.crt TARGET=static HEADLESS=true playwright test tests/inline-menu.spec.ts",
     "test:webserve": "cd clients/apps/web && npm run build:bit:watch",
     "test:webserve:bitwarden": "cd clients/apps/web && ENV=cloud npm run build:oss:watch"
   }
