+++ tests/constants/index.ts
@@ -1,5 +1,5 @@
 export { testPages as autofillTestPages } from "./autofill-forms";
 export { testPages as notificationPages } from "./notification-bar";
-export { testPages as overlayPages } from "./input-overlay";
+export { testPages as inlineMenuPages } from "./inline-menu";
 export { testSiteHost } from "./server";
 export * from "./settings";
